[LayoutAnimations] Fix exiting animations inside a Native Stack Navigator (#3865)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Currently, removing a screen with `exiting` layout animations running
cause views to not be properly removed from the native hierarchy, as
well as leaking memory in Layout Animations' `AnimationManager`.

This is caused by the ancestors of exiting views being removed from the
native hierarchy by the React-Native-Screens implementation, while our
implementation assumes they're still mounted in the native hierarchy.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

In the Example app, go to the "Nested NativeStacks with layout" screen
and try toggling and navigating around while the `exiting` animation on
the second screen of the nested stack is running.