[Android] Fix crash when navigating during an exiting animation (#4002)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->
Fixes a potential crash due to a null pointer exception being thrown.
This crash occurs in the Example app when navigating back during the
exiting animation on the "Nested layout animation" example.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->