[Android] Cleanup animated sensors in NativeProxy destructor (#3954)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
Fixes #3645

The reason for the crash when reloading was that, when the sensor
listeners where cleaned up in the `AnimatedSensorModule`'s destructor,
the `NativeProxy` object used for calling JNI functions has already been
destroyed, causing a null pointer dereference.

This PR moves the sensor listener cleanup to `NativeProxy`'s destructor,
where JNI functions are still safe to call.

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan
In the Example app:
- go to the "Use Animated Sensor" example
- trigger a reload (e.g. by typing "r" in metro)
- 🤞 for no crash

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->