Add web support without a Babel/SWC Plugin (#3997)
## Description

This PR adds support for using Reanimated without a plugin on Web. 

### Before 😢

Without a Babel plugin (or SWC on Next.js), animations do nothing.


https://user-images.githubusercontent.com/13172299/214991524-0819fb9d-d72f-4a2c-b067-3adff7818a50.mp4

### After ✅

Everything now works flawlessly and reactively on Web, including
`useAnimatedStyle`, `useDerivedValue`, and `useAnimatedReaction`.


https://user-images.githubusercontent.com/13172299/214991572-c3a355c2-1a54-43d2-965b-916463b0e281.mp4

The only requirement to make this work is explicitly using dependency
arrays.

## Changes

In places where the `._closure` is undefined, we set the `inputs` to
equal the `dependencies` (so long as they have been explicitly set).

This makes Reanimated work on Web, even if there is no Babel/SWC plugin
set.

Lastly, I added a simple check in `Mapper` to make sure we aren't trying
to get object keys of DOM elements.

## Test code and steps to reproduce

Let me know the best way to do this. I tested the code in a separate
repo outside of this one, and it works well on Web. Native won't be
affected, since all changes are behind a `shouldBeUseWeb()` flag.