[Shared Element Transition] Fix: view disappeared after layout animation (#4026)
## Summary

I added a check to prevent adding already existing views to the shared
transition registry. The issue only occurred when the view had a layout
animation, as the registry is cleared after the animation ends.
Afterwards, views from the registry were removed from their parent view.
In the case of a shared transition, the parent screen is a custom
temporary transition container, but in the case of a regular layout
animation, the super view is the original parent. As a result, the view
disappeared.