Fix perf regression introduced in #3980 🙃 (#4059)
## Summary

This PR replaces implementation of shallowEqual – it turns out that the
new one despite using Object.keys is 20x faster on hermes (haven't
checked JSC). Apparently JIT makes it very hard to measure but the
difference can be easily noticed when running examples with gestures
e.g., snap and drag.

## Test plan

Run snap & drag example on emulator on iOS – with the new implementation
it does not stutter.