Bump react-native to 0.71.3 (#4064)
## Summary

This PR bumps react-native in project root as well as Example and
FabricExample apps from 0.71.1 to 0.71.3.

Although `package.json` previously specified 0.71.1, the app template
was based on 0.71.0, so this PR also contains the changes between .0 and
.1 (see
https://react-native-community.github.io/upgrade-helper/?from=0.70.0&to=0.71.3).

In this PR, I also needed to update plugin tests snapshots because I
upgraded `@babel/*` packages in the project root in order to keep them
aligned with the RN app template. Fortunately, these are only whitespace
changes (empty lines removal).

## Test plan

Run Example and FabricExample apps both on Android and iOS.