Avoid setImmediate calls on Jest environment (#4083)
## Summary

This PR adds Jest specific code to places where we'd used setImmediate.
We want to avoid calling setImmediate as it makes writing tests a bit
tricky.

The reason is that Jest mocks all the timer-related stuff putting all
enqueued function in a single bucket. Then, in order to run these
enqueued functions, the test needs to call advanceTimersByTime. It isn't
allowed to schedule a callback for the same "tick" which results in the
chain of setImmediate calls requiring a several separate runs to
advanceTimersByTime. This can be seen in the previous workaround where
we had three subsequent calls to that method just in order to start a
single animation.

With this change we no longer expect from test to run any additional
setup / timer-advancing steps when testing reanimated components and
animations.

## Test plan

Run jest tests.

---------

Co-authored-by: Tomek Zawadzki <tomasz.zawadzki@swmansion.com>