Fix docs link in reanimated_utils.rb (#4087)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR fixes a broken link in the error message in
`reanimated_utils.rb` which no longer works because `next` version was
removed by @kacperkapusciak in #4066.

## Test plan

Check if the URL is valid.