Hardcode JS library version in `checkVersion` (#4101)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
We hardcode the version of Reanimated here in order to compare it with
the version used to build the native part of the library in runtime.

Ideally, we'd grab this version from the `package.json` file instead,
but this causes issues when packaging the library due to misaligned
relative paths.
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->