Use Node.js 16 on CI (#4274)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR upgrades Node.js from 14 to 16 to match React Native 0.72 (see
[here](https://github.com/facebook/react-native/blob/2c6edee68e936129644c14736e684a14bd11c55c/packages/react-native/template/package.json#L32-L34)).

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->