Change typeof for prototype check to only process plain objects as mappers inputs (#4258)
## Summary

This PR changes the logic on when we scan nested objects provided as
mappers' inputs. Before this change we'd do a `typeof` check to test
whether it is an object. This check however passes on derived objects
such as HostObjects, Maps, etc. Since these types of objects can have
very complex structures we want to avoid scanning them recursively and
only do that for plain object. We achieve this by testing its prototype
and comparing with Object.prototype.

We encountered this error when passing complex objects as style
properties to mappers. Specifically with react-native-skia integration
on
[AnimateTextOnPath](https://github.com/Shopify/react-native-skia/blob/main/example/src/Examples/Reanimated/AnimateTextOnPath.tsx)
example where Path object is used as mappers' input. We only found it
problematic on Web however, because on Native Path object is implemented
as HostObject which in turn does not support Object.values (we always
get empty array for JSI's HostObjects on native). On web, however, the
browser would get stuck on recursively processing this complex object
which was taking very long on Firefox and was crashing the app on
Chrome.

## Test plan

Run test examples in the app
Tested this on web with react-native-skia reanimated AnimateTextOnPath
example where a HostObject is used inside a mapper and it was taking
ages to process that object recursively.